+++ ServiceXpress_Backend/src/main/java/com/backend/controller/ServicePackageController.java
@@ -35,19 +35,19 @@ public class ServicePackageController {
 
     @GetMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasAnyRole('ADMIN', 'CUSTOMER')")
-    public ResponseEntity<ServicePackageDTO> getById(@PathVariable Long id) {
+    public ResponseEntity<ServicePackageDTO> getById(@PathVariable Integer id) {
         return ResponseEntity.ok(packageService.findById(id));
     }
 
     @PutMapping(value = "/{id}", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<ServicePackageDTO> update(@PathVariable Long id, @RequestBody ServicePackageDTO sp) {
+    public ResponseEntity<ServicePackageDTO> update(@PathVariable Integer id, @RequestBody ServicePackageDTO sp) {
         return ResponseEntity.ok(packageService.update(id, sp));
     }
 
     @DeleteMapping("/{id}")
     @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(@PathVariable Integer id) {
         packageService.delete(id);
         return ResponseEntity.noContent().build();
     }
