+++ ServiceXpress_Backend/src/main/java/com/backend/repository/VehicleModelRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VehicleModelRepository extends JpaRepository<VehicleModel, Long> {
+public interface VehicleModelRepository extends JpaRepository<VehicleModel, Integer> {
 }
